Change `origin_main`'s signature to avoid `c_int`.
Change `argc` to `usize` because a common use case is to pass it to
`slice::from_raw_parts`.

Change the return value to `i32`. This matches the type used by
`std::process::exit`.

+++ src/program.rs
@@ -117,7 +117,7 @@ unsafe fn init_runtime(mem: *mut usize, envp: *mut *mut u8) {
 #[allow(unused_variables)]
 unsafe fn call_user_code(argc: c_int, argv: *mut *mut u8, envp: *mut *mut u8) -> i32 {
     extern "Rust" {
-        fn origin_main(argc: c_int, argv: *mut *mut u8, envp: *mut *mut u8) -> c_int;
+        fn origin_main(argc: usize, argv: *mut *mut u8, envp: *mut *mut u8) -> i32;
     }
 
     // Call the functions registered via `.init_array`.
@@ -128,7 +128,7 @@ unsafe fn call_user_code(argc: c_int, argv: *mut *mut u8, envp: *mut *mut u8) ->
     log::trace!("Calling `origin_main({:?}, {:?}, {:?})`", argc, argv, envp);
 
     // Call `origin_main`.
-    let status = origin_main(argc, argv, envp);
+    let status = origin_main(argc as usize, argv, envp);
 
     #[cfg(feature = "log")]
     log::trace!("`origin_main` returned `{:?}`", status);
